// src/api/models/ActualizarClienteNaturalDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ActualizarClienteNaturalDto = {
    nombre: string;
    apellido: string;
    tipoIdentificacion: string;
    numeroIdentificacion: string;
    email: string;
    fechaNacimiento?: string | null;
    genero?: string | null;
    activo?: boolean;
    readonly esValido?: boolean;
};