// src/api/models/CrearClienteEmpresaAdminRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ClienteEmpresaData } from './ClienteEmpresaData';
export type CrearClienteEmpresaAdminRequest = {
    email: string;
    empresa: ClienteEmpresaData;
};