// src/api/models/CrearClienteNaturalDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CrearClienteNaturalDto = {
    nombre: string;
    apellido: string;
    tipoIdentificacion: string;
    numeroIdentificacion: string;
    email: string;
    fechaNacimiento?: string | null;
    genero?: string | null;
    readonly esValido?: boolean;
};